# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HealApp API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/organization/employees": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Get organization employees",
                "description": "Retrieve all employees (users) belonging to the authenticated manager's organization.",
                "operationId": "4aaf5752c7585604673eade6d596653a",
                "responses": {
                    "200": {
                        "description": "Employees retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "first_name": {
                                                "type": "string",
                                                "example": "Maria"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Ivanova"
                                            },
                                            "middle_name": {
                                                "type": "string",
                                                "example": "Petrovna",
                                                "nullable": true
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "9876543210"
                                            },
                                            "organization_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied - user is not a manager",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You must be a manager to access this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/organization/invite-employee": {
            "post": {
                "tags": [
                    "Organization"
                ],
                "summary": "Invite a new employee",
                "description": "Create a new employee (doctor or caregiver) and assign them to the manager's organization. Returns a temporary password for testing.",
                "operationId": "d82ef66bd6ef39f5db306fcc0130d904",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "role"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Employee's phone number (must be unique)",
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "first_name": {
                                        "description": "Employee's first name (optional)",
                                        "type": "string",
                                        "example": "Maria",
                                        "nullable": true
                                    },
                                    "last_name": {
                                        "description": "Employee's last name (optional)",
                                        "type": "string",
                                        "example": "Ivanova",
                                        "nullable": true
                                    },
                                    "middle_name": {
                                        "description": "Employee's middle name",
                                        "type": "string",
                                        "example": "Petrovna",
                                        "nullable": true
                                    },
                                    "role": {
                                        "description": "Employee role",
                                        "type": "string",
                                        "enum": [
                                            "doctor",
                                            "caregiver"
                                        ],
                                        "example": "caregiver"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Maria"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Ivanova"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "9876543210"
                                                },
                                                "organization_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "password": {
                                            "description": "Temporary password for testing (will be sent via SMS in production)",
                                            "type": "string",
                                            "example": "temp123456"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied - user is not a manager",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You must be a manager to invite employees."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}